move start server processing to main
